Added warnings when registering to some shifts or shift templates. Split the shift.views file in several smaller ones.
