+++ tapir/shifts/models.py
@@ -40,6 +40,33 @@ SHIFT_USER_CAPABILITY_CHOICES = {
 }
 
 
+class ShiftSlotWarning:
+    IN_THE_MORNING_EVERYONE_HELPS_STORAGE = "in_the_morning_everyone_helps_storage"
+    IN_THE_EVENING_EVERYONE_HELPS_CLEAN = "in_the_evening_everyone_helps_clean"
+    BREAD_PICKUP_NEEDS_A_VEHICLE = "bread_picked_needs_a_vehicle"
+    MUST_BE_ABLE_TO_CARRY_HEAVY_WEIGHTS = "must_be_able_to_carry_heavy_weights"
+    MUST_NOT_BE_SCARED_OF_HEIGHTS = "must_not_be_scared_of_heights"
+
+
+SHIFT_SLOT_WARNING_CHOICES = {
+    ShiftSlotWarning.IN_THE_MORNING_EVERYONE_HELPS_STORAGE: _(
+        "I understand that all working groups help the Warenannahme & Lagerhaltung working group until the shop opens."
+    ),
+    ShiftSlotWarning.IN_THE_EVENING_EVERYONE_HELPS_CLEAN: _(
+        "I understand that all working groups help the Reinigung & Aufräumen working group after the shop closes."
+    ),
+    ShiftSlotWarning.BREAD_PICKUP_NEEDS_A_VEHICLE: _(
+        "I understand that I need my own vehicle in order to pick up the bread. A cargo bike can be borrowed, more infos in Slack in the #cargobike channel"
+    ),
+    ShiftSlotWarning.MUST_BE_ABLE_TO_CARRY_HEAVY_WEIGHTS: _(
+        "I understand that I may need to carry heavy weights for this shift."
+    ),
+    ShiftSlotWarning.MUST_NOT_BE_SCARED_OF_HEIGHTS: _(
+        "I understand that I may need to work high, for example up a ladder. I do not suffer from fear of heigts."
+    ),
+}
+
+
 class ShiftTemplateGroup(models.Model):
     """ShiftTemplateGroup represents a collection of ShiftTemplates that are usually instantiated together.
 
@@ -295,6 +322,15 @@ class ShiftSlotTemplate(models.Model):
         null=False,
     )
 
+    warnings = ArrayField(
+        models.CharField(
+            max_length=128, choices=SHIFT_SLOT_WARNING_CHOICES.items(), blank=False
+        ),
+        default=list,
+        blank=True,
+        null=False,
+    )
+
     # Whether this ShiftSlot is required to be filled
     optional = models.BooleanField(default=False)
 
@@ -340,6 +376,7 @@ class ShiftSlotTemplate(models.Model):
             name=self.name,
             shift=shift,
             required_capabilities=self.required_capabilities,
+            warnings=self.warnings,
             optional=self.optional,
         )
 
@@ -545,6 +582,15 @@ class ShiftSlot(models.Model):
     required_capabilities = ArrayField(
         models.CharField(
             max_length=128, choices=SHIFT_USER_CAPABILITY_CHOICES.items(), blank=False
+        ),
+        default=list,
+        blank=True,
+        null=False,
+    )
+
+    warnings = ArrayField(
+        models.CharField(
+            max_length=128, choices=SHIFT_SLOT_WARNING_CHOICES.items(), blank=False
         ),
         default=list,
         blank=True,
