+++ tapir/shifts/templates/shifts/shift_day_printable.html
@@ -63,7 +63,7 @@
     <tr>
         <td><h5>#{{ forloop.counter }}</h5></td>
         <td>
-            {{ slot.name }}{% if slot.optional %} ({% translate 'optional' %}){% endif %}<br/>
+            {{ slot.name }}<br/>
             <span class="small-text">
                 {% if slot.required_capabilities|length > 0 %}
                     {% translate 'Required qualifications: ' %}{{ slot.get_required_capabilities_display }}
