--- tapir/shifts/templates/shifts/slot_register.html
@@ -1,46 +1,0 @@
-{% extends "core/base.html" %}
-
-{% load django_bootstrap5 %}
-{% load i18n %}
-{% load static %}
-
-{% block head %}
-    {{ form.media }}
-{% endblock %}
-
-{% block content %}
-    <div class="container">
-        <nav aria-label="breadcrumb">
-            <ol class="breadcrumb">
-                <li class="breadcrumb-item">
-                    {% translate "Shift" %}
-                </li>
-                <li class="breadcrumb-item">
-                    <a href="{{ slot.shift.get_absolute_url }}">
-                        {{ slot.shift.get_display_name }}
-                    </a>
-                </li>
-                <li class="breadcrumb-item">{% translate "Register" %}</li>
-            </ol>
-        </nav>
-
-        <div class="card m-2" id="shift_form_card">
-            <h5 class="card-header d-flex justify-content-between align-items-center">
-                {% translate "Register for Shift" %}:
-                {{ slot.shift.get_display_name }}
-            </h5>
-            <div class="card-body">
-                <form role="form" method="post">
-                    {% csrf_token %}
-                    {% bootstrap_form form %}
-                    <div class="d-flex justify-content-end">
-                        <button type="submit" class="btn tapir-btn btn-primary">
-                            <span class="material-icons">person_add</span>
-                            {% translate 'Register' %}
-                        </button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    </div>
-{% endblock %}
