+++ tapir/shifts/tests/test_member_self_look_for_stand_in.py
@@ -28,7 +28,9 @@ class TestMemberSelfLookForStandIn(TapirSeleniumTestBase):
             self.live_server_url + reverse("shifts:shift_detail", args=[shift.pk])
         )
         self.wait_until_element_present_by_id("shift_detail_card")
-        self.selenium.find_element_by_class_name("register-self-button").click()
+        self.selenium.find_element_by_class_name("register-to-slot-button").click()
+        self.wait_until_element_present_by_id("register_to_shift_slot_card")
+        self.selenium.find_element_by_id("register_button").click()
         self.wait_until_element_present_by_id("self_look_for_stand_in_button")
         self.selenium.find_element_by_id("self_look_for_stand_in_button").click()
         self.wait_until_element_present_by_id("cancel_look_for_stand_in")
@@ -39,7 +41,10 @@ class TestMemberSelfLookForStandIn(TapirSeleniumTestBase):
         self.selenium.get(
             self.live_server_url + reverse("shifts:shift_detail", args=[shift.pk])
         )
-        self.selenium.find_element_by_class_name("register-self-button").click()
+        self.selenium.find_element_by_class_name("register-to-slot-button").click()
+        self.wait_until_element_present_by_id("register_to_shift_slot_card")
+        self.selenium.find_element_by_id("register_button").click()
+
         self.assertEqual(
             ShiftAttendance.objects.filter(
                 user=user2.get_tapir_user(),
