+++ tapir/shifts/views/__init__.py
@@ -1,0 +1,5 @@
+from .attendance import *
+from .calendars import *
+from .exemptions import *
+from .statistics import *
+from .views import *
